Stock count updating correctly. Need a better method for checking if stock exists, currently does not work correctly.

+++ src/test/java/seng202/teamsix/data/StockOperationsStepDefs.java
@@ -10,7 +10,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class StockOperationsStepDefs {
     StockInstance stock_item;
-    float current_stock = 0;
+    double current_stock = 0;
 
     @Given("New stock items arrive")
     public void newStockItemsArrive() {
